requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
flask-cors==4.0.0
bcrypt==4.0.1
cryptography==41.0.3
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
pyfiglet==0.8.post1
colorama==0.4.6
telethon==1.28.5
pytz==2023.3
asyncio
pathlib